--- data/employees.ts
@@ -1,116 +1,0 @@
-import { Employee } from '../types/employee';
-
-export const employees: Employee[] = [
-  {
-    name: 'Airi Satou',
-    position: 'Accountant',
-    office: 'Tokyo',
-    age: 33,
-    startDate: '2008-11-28',
-    salary: 162700,
-  },
-  {
-    name: 'Angelica Ramos',
-    position: 'Chief Executive Officer (CEO)',
-    office: 'London',
-    age: 47,
-    startDate: '2009-10-09',
-    salary: 1200000,
-  },
-  {
-    name: 'Ashton Cox',
-    position: 'Junior Technical Author',
-    office: 'San Francisco',
-    age: 66,
-    startDate: '2009-01-12',
-    salary: 86000,
-  },
-  {
-    name: 'Bradley Greer',
-    position: 'Software Engineer',
-    office: 'London',
-    age: 41,
-    startDate: '2012-10-13',
-    salary: 132000,
-  },
-  {
-    name: 'Brenden Wagner',
-    position: 'Software Engineer',
-    office: 'San Francisco',
-    age: 28,
-    startDate: '2011-06-07',
-    salary: 206850,
-  },
-  {
-    name: 'Brielle Williamson',
-    position: 'Integration Specialist',
-    office: 'New York',
-    age: 61,
-    startDate: '2012-12-02',
-    salary: 372000,
-  },
-  {
-    name: 'Bruno Nash',
-    position: 'Software Engineer',
-    office: 'London',
-    age: 38,
-    startDate: '2011-05-03',
-    salary: 163500,
-  },
-  {
-    name: 'Caesar Vance',
-    position: 'Pre-Sales Support',
-    office: 'New York',
-    age: 21,
-    startDate: '2011-12-12',
-    salary: 106450,
-  },
-  {
-    name: 'Cara Stevens',
-    position: 'Sales Assistant',
-    office: 'New York',
-    age: 46,
-    startDate: '2011-12-06',
-    salary: 145600,
-  },
-  {
-    name: 'Cedric Kelly',
-    position: 'Senior Javascript Developer',
-    office: 'Edinburgh',
-    age: 22,
-    startDate: '2012-03-29',
-    salary: 433060,
-  },
-  {
-    name: 'Charde Marshall',
-    position: 'Regional Director',
-    office: 'San Francisco',
-    age: 36,
-    startDate: '2008-10-14',
-    salary: 470600,
-  },
-  {
-    name: 'Colleen Hurst',
-    position: 'Javascript Developer',
-    office: 'San Francisco',
-    age: 39,
-    startDate: '2009-09-15',
-    salary: 205500,
-  },
-  {
-    name: 'Dai Rios',
-    position: 'Personnel Lead',
-    office: 'Edinburgh',
-    age: 35,
-    startDate: '2012-09-26',
-    salary: 217500,
-  },
-  {
-    name: 'Doris Wilder',
-    position: 'Sales Assistant',
-    office: 'Sydney',
-    age: 23,
-    startDate: '2010-10-20',
-    salary: 85600,
-  },
-];
