--- pages/_app.tsx
@@ -1,12 +1,0 @@
-// pages/_app.js
-// import { CldProvider } from 'next-cloudinary';
-
-// export default function MyApp({ Component, pageProps }) {
-//   return (
-//     <CldProvider
-//       cloudName={process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}
-//     >
-//       <Component {...pageProps} />
-//     </CldProvider>
-//   );
-// }
